+++ sweble-wikitext-components-parent/swc-article-cruncher/src/main/java/org/sweble/wikitext/articlecruncher/pnodes/LpnGatherer.java
@@ -90,7 +90,7 @@ public class LpnGatherer
 						break;
 
 					default:
-						throw new InternalError();
+						throw new AssertionError();
 				}
 
 				processedJobs.put(processedJob);