+++ sweble-wikitext-components-parent/swc-parser-lazy/src/main/java/org/sweble/wikitext/parser/postprocessor/TreeBuilder.java
@@ -205,7 +205,7 @@ public class TreeBuilder
 
 		logic.go(ast);
 		if (getRootNode() == null)
-			throw new InternalError("No root node set after processing!");
+			throw new AssertionError("No root node set after processing!");
 
 		if (!errors.isEmpty())
 		{
@@ -255,7 +255,7 @@ public class TreeBuilder
 			case IN_TABLE_TEXT:
 				return inTableTextMode;
 			default:
-				throw new InternalError();
+				throw new AssertionError();
 		}
 	}
 
@@ -469,7 +469,7 @@ public class TreeBuilder
 					return ((WtLctVarConv) node).getText();
 
 				default:
-					throw new InternalError();
+					throw new AssertionError();
 			}
 		}
 	}
@@ -508,7 +508,7 @@ public class TreeBuilder
 					break;
 
 				default:
-					throw new InternalError();
+					throw new AssertionError();
 			}
 			content.setRtd(originalContent.getRtd());
 		}
@@ -527,9 +527,9 @@ public class TreeBuilder
 		if (nodeType == null)
 		{
 			if (node instanceof WtNamedXmlElement)
-				throw new InternalError("Unknown node type: <" + ((WtNamedXmlElement) node).getName() + ">");
+				throw new AssertionError("Unknown node type: <" + ((WtNamedXmlElement) node).getName() + ">");
 			else
-				throw new InternalError("Unknown node type: " + node.getClass().getSimpleName());
+				throw new AssertionError("Unknown node type: " + node.getClass().getSimpleName());
 		}
 		*/
 		return nodeType;
@@ -641,7 +641,7 @@ public class TreeBuilder
 				((TreeBuilderModeBase) logic.getImpl())
 						.dispatch(getFactory().text(WtRtDataPrinter.print(node)));
 
-				//throw new InternalError();
+				//throw new AssertionError();
 		}
 	}
 
@@ -804,7 +804,7 @@ public class TreeBuilder
 			if (scope.isInList(nodeType))
 				return false;
 		}
-		throw new InternalError("This should never happen!");
+		throw new AssertionError("This should never happen!");
 	}
 
 	boolean isElementTypeInScope(ElementType elementType)
@@ -841,7 +841,7 @@ public class TreeBuilder
 			if (scope.isInList(nodeType))
 				return false;
 		}
-		throw new InternalError("This should never happen!");
+		throw new AssertionError("This should never happen!");
 	}
 
 	boolean isOneOfElementTypesInScope(ElementType... targetTypes)
@@ -895,7 +895,7 @@ public class TreeBuilder
 				return;
 			}
 		}
-		throw new InternalError("Could not remove node from stack!");
+		throw new AssertionError("Could not remove node from stack!");
 	}
 
 	boolean isInStackOfOpenElements(WtNode node)
@@ -926,7 +926,7 @@ public class TreeBuilder
 			if (isSameTag(nodeExample, found))
 				return found;
 		}
-		throw new InternalError("Everything's gone :(");
+		throw new AssertionError("Everything's gone :(");
 	}
 
 	WtNode popFromStackUntilIncluding(ElementType nodeType)
@@ -937,7 +937,7 @@ public class TreeBuilder
 			if (getNodeType(found) == nodeType)
 				return found;
 		}
-		throw new InternalError("Everything's gone :(");
+		throw new AssertionError("Everything's gone :(");
 	}
 
 	WtNode popFromStackUntilIncluding(ElementType... nodeTypes)
@@ -948,7 +948,7 @@ public class TreeBuilder
 			if (isTypeOneOf(getNodeType(found), nodeTypes))
 				return found;
 		}
-		throw new InternalError("Everything's gone :(");
+		throw new AssertionError("Everything's gone :(");
 	}
 
 	void popFromStackUntilIncludingRef(WtNode node)
@@ -956,7 +956,7 @@ public class TreeBuilder
 		while (popFromStack() != node)
 			;
 		if (getStack().isEmpty())
-			throw new InternalError("Everything's gone :(");
+			throw new AssertionError("Everything's gone :(");
 	}
 
 	void popFromStackUntilExcluding(ElementType... nodeTypes)
@@ -964,7 +964,7 @@ public class TreeBuilder
 		while (!isTypeOneOf(getNodeType(getCurrentNode()), nodeTypes))
 			popFromStack();
 		if (getStack().isEmpty())
-			throw new InternalError("Everything's gone :(");
+			throw new AssertionError("Everything's gone :(");
 	}
 
 	WtNode popFromStack()
@@ -1013,7 +1013,7 @@ public class TreeBuilder
 				return;
 			}
 		}
-		throw new InternalError("Marker MUST exist in stack!");
+		throw new AssertionError("Marker MUST exist in stack!");
 	}
 
 	void removeFromParent(WtNode node, WtNode parent)
@@ -1029,7 +1029,7 @@ public class TreeBuilder
 				return;
 			}
 		}
-		throw new InternalError("Node given as parent IS NOT parent of other node!");
+		throw new AssertionError("Node given as parent IS NOT parent of other node!");
 	}
 
 	// =========================================================================
@@ -1162,7 +1162,7 @@ public class TreeBuilder
 				return;
 			}
 		}
-		throw new InternalError("Could not remove formatting element");
+		throw new AssertionError("Could not remove formatting element");
 	}
 
 	void replaceInListOfActiveFormattingElements(
@@ -1178,7 +1178,7 @@ public class TreeBuilder
 				return;
 			}
 		}
-		throw new InternalError("Could not replace formatting element");
+		throw new AssertionError("Could not replace formatting element");
 	}
 
 	/**
@@ -1329,7 +1329,7 @@ public class TreeBuilder
 				return;
 			}
 		}
-		throw new InternalError("This method must only be called if there definitily is a bookmark!");
+		throw new AssertionError("This method must only be called if there definitily is a bookmark!");
 	}
 
 	void moveBookmarkAfter(WtNode node)
@@ -1357,7 +1357,7 @@ public class TreeBuilder
 				return;
 			}
 		}
-		throw new InternalError("This method must only be called if there definitily is a bookmark!");
+		throw new AssertionError("This method must only be called if there definitily is a bookmark!");
 	}
 
 	public void replaceBookmarkWithAndRemove(WtNode replacement, WtNode remove)
@@ -1384,7 +1384,7 @@ public class TreeBuilder
 				return;
 			}
 		}
-		throw new InternalError("This method must only be called if there definitily is a bookmark!");
+		throw new AssertionError("This method must only be called if there definitily is a bookmark!");
 	}
 
 	// =========================================================================