+++ sweble-wikitext-components-parent/swc-parser-lazy/src/main/java/org/sweble/wikitext/parser/postprocessor/TreeBuilderModeBase.java
@@ -148,7 +148,7 @@ public class TreeBuilderModeBase
 
 			default:
 				// Finish is assumed to be WtBold or WtItalics
-				throw new InternalError();
+				throw new AssertionError();
 		}
 	}
 }